BackEnd/requirements.txt
fastapi
uvicorn
sqlmodel
mysqlclient
paho-mqtt
asyncio
pandas